(*) #1095 Moved extensions for SuspensionMode to own class (*) #1095 Added missing comments and remarks
