+++ src/Catel.Core/Catel.Core.Shared/Collections/Interfaces/Enums/SuspensionMode.cs
@@ -38,7 +38,7 @@ namespace Catel.Collections
         /// <summary>
         /// MixedBash mode (combination of Adding and Removing). This behaves the same as <see cref="Mixed"/>, except
         /// that this raises multiply <see cref="NotifyCollectionChangedAction.Add"/> events and <see cref="NotifyCollectionChangedAction.Remove"/>
-        /// ecents instead of single <see cref="NotifyCollectionChangedAction.Reset"/> event.
+        /// events instead of single <see cref="NotifyCollectionChangedAction.Reset"/> event.
         /// </summary>
         MixedBash
     }
