+++ src/Catel.MVVM/Catel.MVVM.Shared/Catel.MVVM.Shared.projitems
@@ -14,6 +14,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Collections\EventArgs\NotifyListChangedEventArgs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Collections\EventArgs\NotifyRangedCollectionChangedEventArgs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Collections\EventArgs\NotifyRangedListChangedEventArgs.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Collections\Extensions\SuspensionModeExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Collections\Extensions\SuspensionContextExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Collections\FastBindingList.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Collections\FastObservableCollection.cs" />
