+++ src/Catel.MVVM/Catel.MVVM.Shared/Collections/EventArgs/NotifyRangedCollectionChangedEventArgs.cs
@@ -22,6 +22,7 @@ namespace Catel.Collections
         /// <summary>
         /// Initializes a new instance of the <see cref="NotifyRangedCollectionChangedEventArgs"/> class.
         /// </summary>
+        /// <remarks>This is only for use of <see cref="Catel.Collections.SuspensionMode.None"/>.</remarks>
         public NotifyRangedCollectionChangedEventArgs()
             : base(NotifyCollectionChangedAction.Reset)
         {
@@ -34,6 +35,7 @@ namespace Catel.Collections
         /// <param name="changedItems">The changed items.</param>
         /// <param name="indices">The indices.</param>
         /// <param name="mode">The suspension mode.</param>
+        /// <remarks>This only for use of <see cref="Catel.Collections.SuspensionMode.Adding"/> and <see cref="Catel.Collections.SuspensionMode.Removing"/>.</remarks>
         public NotifyRangedCollectionChangedEventArgs(IList changedItems, IList<int> indices, SuspensionMode mode)
             : base(ModeToAction(mode), changedItems, (indices != null && indices.Count != 0) ? indices[0] : -1)
         {
@@ -54,6 +56,7 @@ namespace Catel.Collections
         /// <param name="indices">The indices.</param>
         /// <param name="mode">The suspension mode.</param>
         /// <param name="action">The action.</param>
+        /// <remarks>This is only for use of <see cref="Catel.Collections.SuspensionMode.MixedBash"/>.</remarks>
         public NotifyRangedCollectionChangedEventArgs(IList changedItems, IList<int> indices, SuspensionMode mode, NotifyCollectionChangedAction action)
             : base(EnsureModeAndAction(mode, action), changedItems, (indices != null && indices.Count != 0) ? indices[0] : -1)
         {
@@ -73,6 +76,7 @@ namespace Catel.Collections
         /// <param name="changedItems">The changed items.</param>
         /// <param name="indices">The indices.</param>
         /// <param name="mixedActions">The mixed actions.</param>
+        /// <remarks>This is only for use of <see cref="Catel.Collections.SuspensionMode.Mixed"/>.</remarks>
         public NotifyRangedCollectionChangedEventArgs(IList changedItems, IList<int> indices, IList<NotifyCollectionChangedAction> mixedActions)
             : base(NotifyCollectionChangedAction.Reset)
         {
@@ -118,11 +122,13 @@ namespace Catel.Collections
         /// <returns>The <see cref="NotifyCollectionChangedAction"/>.</returns>
         private static NotifyCollectionChangedAction EnsureModeAndAction(SuspensionMode mode, NotifyCollectionChangedAction action)
         {
+            // Check for mixed modes except for Mixed, others fail
             if (mode == SuspensionMode.Mixed || !mode.IsMixedMode())
             {
                 throw new ArgumentException($"Wrong mode '{mode}' for constructor.");
             }
 
+            // Check for action Add or Remove, others fail
             switch (action)
             {
                 case NotifyCollectionChangedAction.Add:
@@ -138,6 +144,7 @@ namespace Catel.Collections
         /// <returns>The equivalent <see cref="NotifyCollectionChangedAction"/>.</returns>
         private static NotifyCollectionChangedAction ModeToAction(SuspensionMode mode)
         {
+            // Only transform modes Adding and Removing, others fail
             switch (mode)
             {
                 case SuspensionMode.Adding: return NotifyCollectionChangedAction.Add;
